Fix implementation of int(str)
A static_cast is required because string to int conversion operator is
marked explicit.

This patch series fixes #2051

+++ pythran/pythonic/numpy/int_.hpp
@@ -24,7 +24,7 @@ namespace numpy
     template <class V>
     long int_(V v)
     {
-      return v;
+      return static_cast<long>(v);
     }
   }
